scraper/requirements.txt
Flask
Werkzeug
requests
beautifulsoup4
datetime
mysql.connector
flask_cors